small additions
